DEVELOPMENT (WORK ONLY)/requirements.txt
google-genai==1.0.0
flask==2.3.3
numpy>=1.23.5,<1.28.0
pillow==9.5.0
pathlib==1.0.1
requests==2.32.3
opencv-python==4.11.0.86
protobuf==4.25.0
pymongo==4.6.1
flask-login==0.6.3
werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.0.1
gunicorn==20.1.0
scipy==1.11.3
scikit-image==0.22.0
# Added for enhanced image analysis
json5==0.9.14
# Added for Stability AI integration
argparse==1.4.0
# Added for StabilityApiGenerator
selenium==4.16.0
webdriver-manager==4.0.1
pyperclip==1.8.2
